+++ app/v0/staking/types/params.go
@@ -149,8 +149,8 @@ func validateUnbondingTime(i interface{}) error {
 		return fmt.Errorf("validateUnbondingTime invalid parameter type: %T", i)
 	}
 
-	if v <= 0 {
-		return fmt.Errorf("unbonding time must be positive: %d", v)
+	if v < 0 {
+		return fmt.Errorf("unbonding time can't be negative: %d", v)
 	}
 
 	return nil