+++ app/v0/staking/types/validator.go
@@ -414,6 +414,9 @@ func (v Validator) BondedTokens() sdk.Int {
 func (v Validator) BondedLever(self_delegation bool, delta sdk.Dec) sdk.Dec {
 	if v.IsBonded() {
 		if self_delegation {
+			if v.SelfDelegation.Add(delta).LTE(sdk.ZeroDec()) {
+				return sdk.NewDec(int64(^uint32(0)))
+			}
 			return v.DelegatorShares.Add(delta).QuoTruncate(v.SelfDelegation.Add(delta))
 		} else {
 			return v.DelegatorShares.Add(delta).QuoTruncate(v.SelfDelegation)